+++ NEWS.md
@@ -1,7 +1,6 @@
 # Changes in version 1.2-0.1 (xx/xx/2016)
 
-* arules_plotly uses now signif instead of round for 
-  quality measures. 
+* arules_plotly uses now signif instead of round for quality measures. 
 * arules_plotly shows correct rule indices for too large rule sets. 
 * arules_plotly implements can now use jitter to reduce overplotting.